feat: separate funding from main flair contract + eslint

+++ contracts/ProxyRegistry.sol
@@ -2,8 +2,8 @@ pragma solidity 0.8.3;
 
 import "@openzeppelin/contracts/access/Ownable.sol";
 
-import "./OwnableDelegateProxy.sol";
-import "./ProxyRegistryInterface.sol";
+import "./lib/proxy/OwnableDelegateProxy.sol";
+import "./lib/proxy/ProxyRegistryInterface.sol";
 
 /**
  * @title ProxyRegistry
@@ -17,7 +17,7 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
     mapping(address => OwnableDelegateProxy) public override proxies;
 
     /* Contracts pending access. */
-    mapping(address => uint) public pending;
+    mapping(address => uint256) public pending;
 
     /* Contracts allowed to call those proxies. */
     mapping(address => bool) public contracts;
@@ -27,7 +27,7 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
        a malicious but rational attacker could buy half the Wyvern and grant themselves access to all the proxy contracts. A delay period renders this attack nonthreatening - given two weeks, if that happened, users would have
        plenty of time to notice and transfer their assets.
     */
-    uint public DELAY_PERIOD = 2 weeks;
+    uint256 public DELAY_PERIOD = 2 weeks;
 
     /**
      * Start the process to enable access for specified contract. Subject to delay period.
@@ -35,10 +35,7 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
      * @dev ProxyRegistry owner only
      * @param addr Address to which to grant permissions
      */
-    function startGrantAuthentication (address addr)
-        public
-        onlyOwner
-    {
+    function startGrantAuthentication(address addr) public onlyOwner {
         require(!contracts[addr] && pending[addr] == 0, "Contract is already allowed in registry, or pending");
         pending[addr] = block.timestamp;
     }
@@ -49,11 +46,11 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
      * @dev ProxyRegistry owner only
      * @param addr Address to which to grant permissions
      */
-    function endGrantAuthentication (address addr)
-        public
-        onlyOwner
-    {
-        require(!contracts[addr] && pending[addr] != 0 && ((pending[addr] + DELAY_PERIOD) < block.timestamp), "Contract is no longer pending or has already been approved by registry");
+    function endGrantAuthentication(address addr) public onlyOwner {
+        require(
+            !contracts[addr] && pending[addr] != 0 && ((pending[addr] + DELAY_PERIOD) < block.timestamp),
+            "Contract is no longer pending or has already been approved by registry"
+        );
         pending[addr] = 0;
         contracts[addr] = true;
     }
@@ -64,10 +61,7 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
      * @dev ProxyRegistry owner only
      * @param addr Address of which to revoke permissions
      */
-    function revokeAuthentication (address addr)
-        public
-        onlyOwner
-    {
+    function revokeAuthentication(address addr) public onlyOwner {
         contracts[addr] = false;
     }
 
@@ -77,10 +71,7 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
      * @dev Must be called by the user which the proxy is for, creates a new AuthenticatedProxy
      * @return proxy New AuthenticatedProxy contract
      */
-    function registerProxy()
-        public
-        returns (OwnableDelegateProxy proxy)
-    {
+    function registerProxy() public returns (OwnableDelegateProxy proxy) {
         return registerProxyFor(msg.sender);
     }
 
@@ -90,11 +81,12 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
      * @dev Must be called by the user which the proxy is for, creates a new AuthenticatedProxy
      * @return proxy New AuthenticatedProxy contract
      */
-    function registerProxyOverride()
-        public
-        returns (OwnableDelegateProxy proxy)
-    {
-        proxy = new OwnableDelegateProxy(msg.sender, delegateProxyImplementation, abi.encodeWithSignature("initialize(address,address)", msg.sender, address(this)));
+    function registerProxyOverride() public returns (OwnableDelegateProxy proxy) {
+        proxy = new OwnableDelegateProxy(
+            msg.sender,
+            delegateProxyImplementation,
+            abi.encodeWithSignature("initialize(address,address)", msg.sender, address(this))
+        );
         proxies[msg.sender] = proxy;
         return proxy;
     }
@@ -105,12 +97,13 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
      * @dev Can be called by any user
      * @return proxy New AuthenticatedProxy contract
      */
-    function registerProxyFor(address user)
-        public
-        returns (OwnableDelegateProxy proxy)
-    {
+    function registerProxyFor(address user) public returns (OwnableDelegateProxy proxy) {
         require(address(proxies[user]) == address(0), "User already has a proxy");
-        proxy = new OwnableDelegateProxy(user, delegateProxyImplementation, abi.encodeWithSignature("initialize(address,address)", user, address(this)));
+        proxy = new OwnableDelegateProxy(
+            user,
+            delegateProxyImplementation,
+            abi.encodeWithSignature("initialize(address,address)", user, address(this))
+        );
         proxies[user] = proxy;
         return proxy;
     }
@@ -118,9 +111,7 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
     /**
      * Transfer access
      */
-    function transferAccessTo(address from, address to)
-        public
-    {
+    function transferAccessTo(address from, address to) public {
         OwnableDelegateProxy proxy = proxies[from];
 
         /* CHECKS */
@@ -131,5 +122,4 @@ contract ProxyRegistry is Ownable, ProxyRegistryInterface {
         delete proxies[from];
         proxies[to] = proxy;
     }
-
 }
